Выносит рендер списка с фильмами в отдельный файл
